requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
torch>=1.10.0
python-dotenv>=0.19.0
openai>=0.27.0
azure-identity>=1.10.0
loguru>=0.6.0
